Revert "Help Center History: Changed date format (#100287)"
This reverts commit 895bf5cf0f045fced3b4ee251cb281a6eb5e24de.
