+++ client/landing/subscriptions/components/site-subscriptions-list/site-subscription-row.tsx
@@ -1,4 +1,4 @@
-import { Gridicon, ExternalLink, TimeSince } from '@automattic/components';
+import { Gridicon, ExternalLink } from '@automattic/components';
 import { Reader, SubscriptionManager } from '@automattic/data-stores';
 import { localizeUrl } from '@automattic/i18n-utils';
 import { __experimentalHStack as HStack } from '@wordpress/components';
@@ -7,6 +7,7 @@ import { useMemo, useRef } from 'react';
 import { useDispatch } from 'react-redux';
 import { SiteIcon } from 'calypso/blocks/site-icon';
 import InfoPopover from 'calypso/components/info-popover';
+import TimeSince from 'calypso/components/time-since';
 import {
 	useRecordSiteUnsubscribed,
 	useRecordSiteResubscribed,
