+++ client/my-sites/activity/activity-log/threat-alert.jsx
@@ -1,4 +1,4 @@
-import { Button, Card, Spinner, FoldableCard, TimeSince } from '@automattic/components';
+import { Button, Card, Spinner, FoldableCard } from '@automattic/components';
 import { localizeUrl } from '@automattic/i18n-utils';
 import { JETPACK_CONTACT_SUPPORT } from '@automattic/urls';
 import clsx from 'clsx';
@@ -11,6 +11,7 @@ import ThreatDialog from 'calypso/components/jetpack/threat-dialog';
 import MarkedLines from 'calypso/components/marked-lines';
 import PopoverMenuItem from 'calypso/components/popover-menu/item';
 import SplitButton from 'calypso/components/split-button';
+import TimeSince from 'calypso/components/time-since';
 import { Interval, EVERY_TEN_SECONDS } from 'calypso/lib/interval';
 import { recordTracksEvent, withAnalytics } from 'calypso/state/analytics/actions';
 import { fixThreatAlert, ignoreThreatAlert } from 'calypso/state/jetpack/site-alerts/actions';
