+++ client/my-sites/earn/customers/customer/customer-details.tsx
@@ -1,5 +1,5 @@
-import { TimeSince } from '@automattic/components';
 import { useTranslate, formatCurrency } from 'i18n-calypso';
+import TimeSince from 'calypso/components/time-since';
 import {
 	PLAN_YEARLY_FREQUENCY,
 	PLAN_MONTHLY_FREQUENCY,
