+++ packages/help-center/src/components/help-center-support-chat-message.tsx
@@ -1,7 +1,6 @@
-/* eslint-disable no-restricted-imports */
 import { recordTracksEvent } from '@automattic/calypso-analytics';
-import { Gravatar, TimeSince } from '@automattic/components';
-import { getNumericDateString, useLocale } from '@automattic/i18n-utils';
+import { Gravatar } from '@automattic/components';
+import { getRelativeTimeString, useLocale } from '@automattic/i18n-utils';
 import { HumanAvatar } from '@automattic/odie-client/src/assets';
 import { useDispatch as useDataStoreDispatch } from '@wordpress/data';
 import { chevronRight, Icon } from '@wordpress/icons';
@@ -13,6 +12,7 @@ import { HELP_CENTER_STORE } from '../stores';
 import type { SupportInteraction, ZendeskMessage } from '@automattic/odie-client';
 
 import './help-center-support-chat-message.scss';
+
 const trackContactButtonClicked = ( sectionName: string ) => {
 	recordTracksEvent( 'calypso_inlinehelp_support_chat_message_click', {
 		force_site_id: true,
@@ -108,7 +108,11 @@ export const HelpCenterSupportChatMessage = ( {
 						}
 					/>
 					<span>
-						<TimeSince date={ getNumericDateString( received * 1000, locale ) } />
+						{ getRelativeTimeString( {
+							timestamp: received * 1000,
+							locale,
+							style: 'long',
+						} ) }
 					</span>
 				</div>
 			</div>
